The ugliest commit yet
decided to serialize each output with TIGGER Serial number, inject it with a P4TIGTAG: HRA, SUPPORT, ALL. Then build a json file out of it for ease of breaking that file down into several files for delegation. 

Ultimately I'd like to send a json to the datapushgateway and break the json down their for parsing.